requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
requests==2.32.3
pywebview==5.3.2
pyinstaller==6.11.1
